フロントエンド、class化 - UI操作の自動テスト #6

+++ static/js/app.js
@@ -1,28 +1,68 @@
-function starButton(postId, userId, starId=0) {
-  const csrftoken = document.querySelector('[name=csrfmiddlewaretoken]').value;
-  const _postId = postId;
-  const _userId = userId;
-  let _starId = starId;
-  const element = document.querySelector(`.star.star--post-id-${_postId}`);
-  if (_starId) {
-    element.addEventListener('click', removeStar)
-  } else {
-    element.addEventListener('click', addStar)
-  }
+class starButton {
+  constructor(csrftoken, postId, userId, starId=0) {
+    this.csrftoken = csrftoken;
+    this.postId = postId;
+    this.userId = userId;
+    this.starId = starId;
+    this.element = document.querySelector(`.star.star--post-id-${postId}`);
 
-  function addStar() {
-    fetch("/api/stars/",{
-      method: "POST",
-      headers: {
-        'X-CSRFToken': csrftoken,
-        'Content-Type': 'application/json'
-      },
-      body:JSON.stringify({
-          post: _postId,
-          user: _userId
-      })
-    })
-    .then(function(response) {
+    this.isFetching = false;
+
+    /**
+     * 通信中にいいね!ボタン押下不可にする
+     */
+    this.disabled = () => {
+      this.element.style.pointerEvents = 'none';
+      this.isFetching = true;
+    }
+
+    /**
+     * いいね!ボタン押下可能にする
+     */
+    this.enabled = () => {
+      this.element.style.pointerEvents = '';
+      this.isFetching = false;
+    }
+
+    /**
+     * 初期化
+     */
+    this.init = () => {
+      if (this.starId) {
+        // いいね済の場合
+        this.element.textContent = 'いいね済';
+        this.element.removeEventListener('click', this.addStar)
+        this.element.addEventListener('click', this.removeStar)
+      } else {
+        // まだいいね!していない
+        this.element.textContent = 'いいねする';
+        this.element.removeEventListener('click', this.removeStar)
+        this.element.addEventListener('click', this.addStar)
+      }
+    }
+
+    /**
+     * 非同期通信でいいね!する
+     */
+    this.addStar = async () => {
+      // 連打対策
+      if (this.isFetching) {
+        console.log('通信中のためskip');
+        return;
+      }
+      this.disabled();
+
+      let response = await fetch("/api/stars/", {
+        method: "POST",
+        headers: {
+          'X-CSRFToken': this.csrftoken,
+          'Content-Type': 'application/json'
+        },
+        body:JSON.stringify({
+            post: this.postId,
+            user: this.userId
+        })
+      });
       if (!response.ok) {
         throw new Error('Network response was not ok');
       }
@@ -30,43 +70,40 @@ function starButton(postId, userId, starId=0) {
       if (response.status !== 201) {
         throw new Error(response.status);
       }
-      element.textContent = 'いいね済';
-      return response.json();
-    })
-    .then(function(data) {
-      console.log(JSON.stringify(data));
-      element.addEventListener('click', removeStar)
-      element.removeEventListener('click', addStar)
-      _starId = data['id']
-    })
-    .catch(function(err) {
-      console.log(err);
-    });
-  }
+      const data = await response.json();
+      this.starId = data['id']
+      console.log(data);
+      this.init();
+      this.enabled();
+    }
 
-  function removeStar() {
-    fetch(`/api/stars/${_starId}`, {
-      method: "DELETE",
-      headers: {
-        'X-CSRFToken': csrftoken,
-        'Content-Type': 'application/json'
+    /**
+     * 非同期通信でいいね!を取り消す
+     */
+    this.removeStar = async () => {
+      // 連打対策
+      if (this.isFetching) {
+        console.log('通信中のためskip');
+        return;
       }
-    })
-    .then(function(response) {
+      this.disabled();
+
+      let response = await fetch(`/api/stars/${this.starId}`, {
+        method: "DELETE",
+        headers: {
+          'X-CSRFToken': this.csrftoken,
+          'Content-Type': 'application/json'
+        }
+      })
       if (!response.ok) {
         throw new Error('Network response was not ok');
       }
-      console.log("status=" + response.status);
       if (response.status !== 204) {
         throw new Error(response.status);
       }
-      _starId = 0;
-      element.textContent = 'いいねする';
-      element.addEventListener('click', addStar)
-      element.removeEventListener('click', removeStar)
-    })
-    .catch(function(err) {
-      console.log(err);
-    });
+      this.starId = 0;
+      this.init();
+      this.enabled();
+    }
   }
 }
